refactor(core): delete the orphaned aidlc package (zero production callers)
src/kiro_crew/aidlc/ (models.py, store.py, __init__.py) has no production
consumer: no module under src/ imports it outside the package itself, no
dynamic import by string name exists, and setup.cfg discovers packages via
find: so no manifest enumerates it. The only references were its own test
(test/test_aidlc_store.py) and the dead-subtree exclusion entry in
test/test_pip_deps_consistency.py.

Maintainer decision on the issue: delete, not wire up. This removes the
package, its test, and the now-dead exclusion entry (docstring + the
"aidlc/" line in _EXCLUDED_SUBTREES) without restructuring the
deps-consistency gate.

The deleted test file's TestEnrichedDataclasses class covered live
kiro_crew.task_models fields (priority, story_points, task_type, mode,
source_spec, skip_planning), not the deleted store - those assertions are
preserved in the new test/test_task_models_enriched.py rather than dying
with the package.

The website/src/pages/aidlc/ frontend pages are unrelated: they import only
within website/ and are served by handlers_project.py off the task runner,
not this store. They are untouched.

Closes #4528
