--- src/kiro_crew/aidlc/__init__.py
@@ -1,1 +1,0 @@
-"""Project management store — CRUD for projects, activities, comments."""
